fix: rework php file structure
