+++ src/Core/Assets.php
@@ -7,10 +7,10 @@
 
 namespace SesamyPlugin\Core;
 
-use function SesamyPlugin\Support\Helpers\get_enabled_post_types;
-use function SesamyPlugin\Support\Helpers\is_config_valid;
-use function SesamyPlugin\Support\Helpers\get_sesamy_setting;
-use function SesamyPlugin\Support\Helpers\is_update_available;
+use function SesamyPlugin\Helpers\get_enabled_post_types;
+use function SesamyPlugin\Helpers\is_config_valid;
+use function SesamyPlugin\Helpers\get_sesamy_setting;
+use function SesamyPlugin\Releases\is_update_available;
 
 /**
  * Assets module.
